small changes / refactoring
