+++ rvas-core_v0.3/pom.xml
@@ -75,7 +75,7 @@
     </dependencies>
 
     <build>
-        <finalName>rvas-core-0.3.3-289</finalName>
+        <finalName>rvas-core-0.3.3-290</finalName>
         <plugins>
             <!-- download source code in Eclipse, best practice-->
             <plugin>
