+++ rvas-core_v0.3/src/main/java/core/commands/restricted/Info.java
@@ -1,4 +1,4 @@
-package core.commands;
+package core.commands.restricted;
 
 /* *
  *
